feat: authenticate streaming with OAuth.
It looks like Tidal uses OAuth to retrieve a Bearer token both for the
API and, as one would expect, for normal users. We can steal the client
ID and secret off of the network inspector. These are of course
longer-lived credentials than the session token, so they should be more
convenient to use.

(I still can't think of a good way to authenticate just via
username/password. This is fine for now.)